CI: enable `dune-cache` (#3)
I don't think this will cut down on all of the CI build time issues (esp. since much of it is the tooling to build the compiler itself), but this should help. Only concern is if the cache has issues re: bumping versions of RSDD, but I'll keep that in mind (and revert this when necessary).